Back out 902a7c3eb320 for failing to build on Windows, and 196df669baba for failing to build without it
UltraBlame original commit: 54a15fc04437f437a88253f971e343b03f591b0a

+++ xpcom/ds/nsCheapSets.h
@@ -39,7 +39,6 @@
 #define __nsCheapSets_h__
 
 #include "nsHashSets.h"
-#include <stdint.h>
 
 
 
@@ -176,7 +175,7 @@ private:
   
   void SetInt(PRInt32 aInt)
   {
-    mValOrHash = (void*)(intptr_t)((aInt << 1) | 0x1);
+    mValOrHash = (void*)((aInt << 1) | 0x1);
   }
   
   nsresult InitHash(nsInt32HashSet** aSet);
